requirements.txt
Package                                  Version
---------------------------------------- -----------
aenum                                    3.1.15
aiohappyeyeballs                         2.4.6
aiohttp                                  3.11.12
aiosignal                                1.3.2
annotated-types                          0.7.0
anyio                                    4.8.0
asgiref                                  3.8.1
attrs                                    25.1.0
certifi                                  2025.1.31
charset-normalizer                       3.4.1
click                                    8.1.8
Deprecated                               1.2.18
fastapi                                  0.115.8
frozenlist                               1.5.0
future                                   1.0.0
googleapis-common-protos                 1.66.0
grpcio                                   1.70.0
h11                                      0.14.0
idna                                     3.10
importlib_metadata                       8.5.0
line-bot-sdk                             3.14.5
multidict                                6.1.0
opentelemetry-api                        1.30.0
opentelemetry-exporter-otlp              1.30.0
opentelemetry-exporter-otlp-proto-common 1.30.0
opentelemetry-exporter-otlp-proto-grpc   1.30.0
opentelemetry-exporter-otlp-proto-http   1.30.0
opentelemetry-instrumentation            0.51b0
opentelemetry-instrumentation-asgi       0.51b0
opentelemetry-instrumentation-fastapi    0.51b0
opentelemetry-proto                      1.30.0
opentelemetry-sdk                        1.30.0
opentelemetry-semantic-conventions       0.51b0
opentelemetry-util-http                  0.51b0
packaging                                24.2
pip                                      25.0
propcache                                0.2.1
protobuf                                 5.29.3
pydantic                                 2.10.6
pydantic_core                            2.27.2
pydantic-settings                        2.7.1
python-dateutil                          2.9.0.post0
python-dotenv                            1.0.1
python-json-logger                       3.2.1
PyYAML                                   6.0.2
requests                                 2.32.3
setuptools                               75.8.0
six                                      1.17.0
sniffio                                  1.3.1
starlette                                0.45.3
typing_extensions                        4.12.2
urllib3                                  2.3.0
uvicorn                                  0.34.0
wheel                                    0.45.1
wrapt                                    1.17.2
yarl                                     1.18.3
zipp                                     3.21.0
